Splits parse and merging functions to prepare for async introduction
By splitting the parsing from the merging it will be easier to leave any
async handling out of the merging functionality. Async functions are only
necessary for the loading of remote files but not for the merging
itself. This will keep the functions with clearer responsibilities as
well as makes testing simpler.