Moves function bodies from Processes impl to docker module
Not sure if it really pays off, but at the moment it feels like this
should be abstracted in its own "docker" module, so that we could
eventually test these functions in isolation. Plus not having the
intricacies of how to interact with Docker inside the processes module
seems to be a worthwile benefit.